requirements.txt
Django>=2.2.0,<3
django-nose
coverage
djangorestframework>=3.11.0,<3.12.0
djangorestframework-camel-case>=1.2.0,<1.4.0
black
argon2-cffi
drf-yasg>=1.17,<1.19
